pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crewai-ollama-stack"
version = "0.1.0"
description = "Self-contained CrewAI + Ollama deployment tooling with crewctl CLI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "CrewAI Maintainers" }]
dependencies = [
  "typer>=0.12.3,<0.13",
  "rich>=13.7.0,<14",
  "pyyaml>=6.0.1,<7",
]
keywords = ["crewai", "ollama", "deployment", "cli"]

[project.optional-dependencies]
train = [
  "torch>=2.2.0",
  "transformers>=4.38.0",
  "datasets>=2.16.0",
  "peft>=0.9.0",
  "accelerate>=0.27.0",
  "sentencepiece>=0.1.99",
  "scikit-learn>=1.3.0",
  "bitsandbytes>=0.43.0; platform_system != 'Windows'",
]

[project.scripts]
crewctl = "crewctl.__main__:main"

[tool.setuptools]
packages = ["crewctl"]

[tool.setuptools.package-data]
crewctl = ["templates/*"]
